fix: delete subs err

+++ routes/accounts/[id]/subscriptions/index.delete.ts
@@ -4,6 +4,6 @@ export default defineAuthenticatedEventHandler(async (event) => {
   const { id } = getRouterParams(event)
   const subscriptions = await prisma.accountSubscription.findMany({ where: { accountId: id } })
   const servers = await prisma.server.findMany({ where: { OR: subscriptions.map(({ subscriptionId }) => ({ subscriptionId })) } })
-  await prisma.accountServer.deleteMany({ where: { OR: servers.map(({ id }) => ({ serverId: id })) } })
+  await prisma.accountServer.deleteMany({ where: { accountId: id, OR: servers.map(({ id }) => ({ serverId: id })) } })
   return prisma.accountSubscription.deleteMany({ where: { accountId: id } })
 })
